requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aioschedule==0.5.2
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
frozenlist==1.6.0
google-auth==2.40.2
google-auth-oauthlib==1.2.2
greenlet==3.2.2
gspread==6.2.1
httplib2==0.22.0
idna==3.10
isodate==0.7.2
lxml==5.4.0
magic-filter==1.0.12
multidict==6.4.4
oauth2client==4.1.3
oauthlib==3.2.2
platformdirs==4.3.8
propcache==0.3.1
py-dotenv==0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
PyMySQL==1.1.1
pyparsing==3.2.3
python-dotenv==1.1.0
pytz==2025.2
redis==6.1.0
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9.1
schedule==1.2.2
six==1.17.0
SQLAlchemy==2.0.41
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0
zeep==4.3.1
